feat(user-service): convert package to esm
